+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - new events `import-data` and `import-data/user-config` are available
+- `patchMultipart` implementation for the http client
 
 ## [1.0.10] - 2023-06-22